+++ CHANGELOG.md
@@ -14,6 +14,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.0.3] - 2025-08-01
+Added:
+- generation of reddit posts
+
+Changed:
+- i no longer create tweet for all summaries, only the ones i select
+
+Fixed:
+- write tweet as a person
+
+## [0.0.3] - 2025-08-01
 - fix the layout of the post card
 
 ## [0.0.2] - 2025-07-31